INTGR-142: refactor/Rename Hardware Code Samples sub-folder  - Rename from 'samples' to 'codeSamples'
