+++ src/pages/hardware/codeSamples/python.tsx
